# plugin-kit/samples/openai-gpt-plugin/src/controllers/config.ts
import { Request, Response } from 'express';
import { MetricsService } from '../services/metrics';
import { LoggerService } from '../services/logger';

export class ConfigController {
  constructor(
    private metrics: MetricsService,
    private logger: LoggerService
  ) {}

  async getConfig(req: Request, res: Response): Promise<void> {
    try {
      const config = {
        plugin: {
          name: 'OpenAI GPT Plugin',
          version: process.env.npm_package_version || '1.0.0',
          description: 'AI-powered code completion and analysis plugin using OpenAI GPT',
          author: 'CircuitExp1',
          license: 'MIT',
        },
        features: {
          completion: true,
          review: true,
          analysis: {
            performance: true,
            security: true,
            complexity: true,
            maintainability: true,
          },
          bugDetection: true,
          testGeneration: true,
          documentation: true,
          chat: true,
        },
        models: {
          default: 'gpt-4',
          available: ['gpt-4', 'gpt-4-turbo', 'gpt-3.5-turbo', 'gpt-3.5-turbo-16k'],
        },
        limits: {
          maxTokens: {
            completion: 2048,
            review: 4096,
            analysis: 2048,
            chat: 2048,
          },
          maxCodeLength: 50000,
          maxContextLength: 10000,
        },
        endpoints: [
          'GET /health',
          'GET /ready',
          'GET /live',
          'POST /api/completion',
          'POST /api/completion/stream',
          'POST /api/review',
          'POST /api/review/quick',
          'POST /api/analysis',
          'POST /api/analysis/bugs',
          'POST /api/analysis/tests',
          'POST /api/analysis/docs',
          'POST /api/chat',
          'POST /api/chat/stream',
          'GET /api/config',
          'POST /api/config',
        ],
      };

      this.logger.debug('Configuration requested');
      res.json({
        success: true,
        data: config,
      });
    } catch (error) {
      this.logger.error('Failed to get configuration', error);
      res.status(500).json({
        success: false,
        error: 'Failed to retrieve configuration',
        message: error instanceof Error ? error.message : 'Unknown error',
      });
    }
  }

  async updateConfig(req: Request, res: Response): Promise<void> {
    try {
      // Note: In a real implementation, this would update configuration
      // For now, we'll return a not implemented response

      this.logger.warn('Configuration update requested - not implemented');
      res.status(501).json({
        success: false,
        error: 'Configuration updates not implemented',
        message: 'Configuration updates are not supported in this version',
      });
    } catch (error) {
      this.logger.error('Failed to update configuration', error);
      res.status(500).json({
        success: false,
        error: 'Failed to update configuration',
        message: error instanceof Error ? error.message : 'Unknown error',
      });
    }
  }
}